requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
attrs==21.2.0
    # via pytest
beautifulsoup4==4.10.0
    # via -r requirements.in
cachecontrol==0.12.10
    # via firebase-admin
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via flask
firebase-admin==5.1.0
    # via -r requirements.in
flask==2.0.2
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==3.0.10
    # via -r requirements.in
google-api-core[grpc]==2.2.2
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-logging
    #   google-cloud-storage
google-api-python-client==2.32.0
    # via firebase-admin
google-auth==2.3.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-appengine-logging==1.1.0
    # via google-cloud-logging
google-cloud-audit-log==0.2.0
    # via google-cloud-logging
google-cloud-core==2.2.1
    # via
    #   google-cloud-firestore
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-firestore==2.3.4
    # via firebase-admin
google-cloud-logging==2.7.0
    # via -r requirements.in
google-cloud-storage==1.43.0
    # via firebase-admin
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.1.0
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.53.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.3
    # via google-cloud-logging
grpcio==1.42.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.42.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.20.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
itsdangerous==2.0.1
    # via flask
jinja2==3.0.3
    # via flask
markupsafe==2.0.1
    # via jinja2
msgpack==1.0.3
    # via cachecontrol
packaging==21.3
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-firestore
    #   pytest
pluggy==1.0.0
    # via pytest
proto-plus==1.19.8
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-firestore
    #   google-cloud-logging
protobuf==3.19.1
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   google-cloud-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.6
    # via
    #   httplib2
    #   packaging
pytest==6.2.5
    # via -r requirements.in
requests==2.26.0
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   flask-cors
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-storage
    #   grpcio
soupsieve==2.3.1
    # via beautifulsoup4
toml==0.10.2
    # via pytest
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.7
    # via requests
werkzeug==2.0.2
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
